backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
cryptography==42.0.5
pg8000==1.31.2
python-dotenv==1.0.0
# psycopg2-binary==2.9.9  # Replaced with pg8000 due to Python 3.13 compatibility
PyPDF2==3.0.1
bcrypt==4.1.2
email-validator==2.1.0
itsdangerous==2.1.2
